Feat: Spit up the mix task into 2 modules
Problem:
- current module is too hard to use from the IEx shell
- also it hides most features behind private functions
- this makes reuse or unit testing harder

Solution:
- split up into 2 modules
- make features accessible
- do not write files by default
